Adds Create Camel project to the welcome screen
- Rename NewCamelProjectCommand class to reflect its abstract nature.
- Create a 'generic' NewCamelProjectCommand that asks the user to select
  a runtime. Register the command with 'enablement = false' so that its
  does not show in the command palette.
- Refactor other classes with the new project structure.
